Add tracexec - flush/grab/ungrab evdev other proc
ptrace() into tracee, and allow arbitrary assembly instruction
to executed inside the tracee. comes with convenience wrappers
for amd64 and arm syscalls. By default, it allows to
drain a evdev event stream while a program is paused,
or to ungrab a grabbed evdev file descriptor,
but it should be easy to adopt it to do different things.

Very hacky, and very much GNU extensions tied.