Add keyboard http server (untested)
inkvt will now listen on port 7800 for keyboard input,
which will be sent directly to pty input.
The http server is absolutely not production ready, and
blocks the entire main thread while processing data (which
might even hang forever).
See kbsend.html for a usage example (which should work on
local host from the browser).
The goal is to be able to use a smartphone as an external
keyboard.