Nickle/KOReader takeover - behind -DEXPERIMENTAL
It should work in theory - but does not work in
practice. Still figuring out why.

Testing is possible with the -DEXPERIMENTAL flag,
or by running tracexec.x, but very much not
recommendend on arm devices.

tracexec.x86 is a test binary that simply lists
all file descriptor flags of the program with the
pid given on the command line, by using syscalls inside
the tracee.

```
(terminal one)
$ cat

(terminal two)
$ make build/tracexec.x
$ sudo ./build/tracexec.x86 $(pidof cat)
[sudo] password for llandsmeer:
[0]: 2
[1]: 2
[2]: 2
```