tsyscall
Got rid of old tracexec. Rewrote the entire syscall
in tracee thing - this time with an actual arm emulator.

It. Still. Does. Not. Work.

ptrace() problems on arm:
 - if the tracee is waiting on a syscall result on PTRACE_SEIZE,
   strange things happen. that is kindof expected, but wasn't
   a problem on my amd64 laptop..
 - the process might be in thumb mode. or it might not be.
   at them moment I just zero the Thumb mode cpsr flag,
   which seems to work.
 - PTRACE_SINGLESTEP is apparantly not implemented

So I spend way too much time on this already. It works in an arm
emulator and it works on my laptop. Just not the kobo. The
development iteration time is just way too high to debug this
stupid problem. I'll just go with the koreader shell scripts.

However, I got the kobo to list the filedescriptor flags of `cat`,
from within `cat`. Thats about it.